backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysql-connector-python
pydantic
websockets
python-decouple 